Core data objects refactor (#66)
* Refactored Settings

Removed dailyResetWarning
Modified vars to remove outdated CoreData artifacts

* Int64 -> Int

* Adjust Quest vars

* Update User.swift

* made init vars more clear

* Removed superfluous Int casting

* merge straggler

+++ Quest Tracker/Rewards/AddRewardView.swift
@@ -43,9 +43,9 @@ struct AddRewardView: View {
     .onDisappear(perform: {
       if reward.name.count > 0 {
         if reward.isMilestoneReward {
-          reward.sortId = Int64(milestoneRewardCount)
+          reward.sortId = milestoneRewardCount
         } else {
-          reward.sortId = Int64(minorRewardCount)
+          reward.sortId = minorRewardCount
         }
 
         modelContext.insert(reward)
